Improvy copy and move parts of the code to different files
